creating black duck config
