+++ src/main/java/com/synopsys/integration/alert/ApplicationConfiguration.java
@@ -13,6 +13,7 @@ import org.springframework.batch.core.launch.support.SimpleJobLauncher;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.repository.support.MapJobRepositoryFactoryBean;
 import org.springframework.batch.support.transaction.ResourcelessTransactionManager;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -25,25 +26,24 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.csrf.HttpSessionCsrfTokenRepository;
 import org.springframework.transaction.PlatformTransactionManager;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.synopsys.integration.alert.common.AlertProperties;
 import com.synopsys.integration.alert.common.descriptor.accessor.RoleAccessor;
 import com.synopsys.integration.alert.common.persistence.util.FilePersistenceUtil;
 import com.synopsys.integration.alert.common.security.EncryptionUtility;
 import com.synopsys.integration.alert.common.security.authorization.AuthorizationManager;
-import com.synopsys.integration.blackduck.http.transform.BlackDuckJsonTransformer;
-import com.synopsys.integration.blackduck.http.transform.subclass.BlackDuckResponseResolver;
-import com.synopsys.integration.blackduck.service.BlackDuckServicesFactory;
-import com.synopsys.integration.log.Slf4jIntLogger;
-import com.synopsys.integration.rest.RestConstants;
-import com.synopsys.integration.rest.support.AuthenticationSupport;
 
 @Configuration
 @AutoConfigureOrder(1)
 public class ApplicationConfiguration {
     private final Logger logger = LoggerFactory.getLogger(ApplicationConfiguration.class);
+
+    private final Gson gson;
+
+    @Autowired
+    public ApplicationConfiguration(Gson gson) {
+        this.gson = gson;
+    }
 
     @Bean
     public AlertProperties alertProperties() {
@@ -52,7 +52,7 @@ public class ApplicationConfiguration {
 
     @Bean
     public FilePersistenceUtil filePersistenceUtil() {
-        return new FilePersistenceUtil(alertProperties(), gson());
+        return new FilePersistenceUtil(alertProperties(), gson);
     }
 
     @Bean
@@ -99,31 +99,6 @@ public class ApplicationConfiguration {
     @Bean
     public TaskExecutor taskExecutor() {
         return new SyncTaskExecutor();
-    }
-
-    @Bean
-    public Gson gson() {
-        return BlackDuckServicesFactory.createDefaultGson();
-    }
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        return BlackDuckServicesFactory.createDefaultObjectMapper();
-    }
-
-    @Bean
-    public BlackDuckResponseResolver blackDuckResponseResolver() {
-        return new BlackDuckResponseResolver(gson());
-    }
-
-    @Bean
-    public BlackDuckJsonTransformer blackDuckJsonTransformer() {
-        return new BlackDuckJsonTransformer(gson(), objectMapper(), blackDuckResponseResolver(), new Slf4jIntLogger(logger));
-    }
-
-    @Bean
-    public AuthenticationSupport authenticationSupport() {
-        return new AuthenticationSupport();
     }
 
     @Bean
